requirements.txt
dashscope==1.15.0
matplotlib==3.8.3
pillow==10.2.0
numpy==1.26.4
openai==1.52.0
pandas==1.5.3
sqlalchemy==2.0.29
pyyaml==6.0.1
uvicorn~=0.31.0
fastapi~=0.115.0
pydantic~=2.9.2
pyecharts==2.0.7
pygwalker==0.4.8.7
pywebio== 1.8.3
PyMySQL==1.1.1
httpx==0.27.2